+++ apispec/ext/marshmallow/swagger.py
@@ -20,6 +20,10 @@ from apispec.lazy_dict import LazyDict
 
 ##### marshmallow #####
 
+MARSHMALLOW_VERSION_INFO = tuple(
+    [int(part) for part in marshmallow.__version__.split('.') if part.isdigit()]
+)
+
 # marshmallow field => (JSON Schema type, format)
 FIELD_MAPPING = {
     marshmallow.fields.Integer: ('integer', 'int32'),
@@ -89,12 +93,13 @@ def _observed_name(field, name):
     :param str name: Field name
     :rtype: str
     """
-    # use getattr in case we're running against older versions of marshmallow.
-    dump_to = getattr(field, 'dump_to', None)
-    load_from = getattr(field, 'load_from', None)
-    if load_from != dump_to:
-        return name
-    return dump_to or name
+    if MARSHMALLOW_VERSION_INFO[0] < 3:
+        # use getattr in case we're running against older versions of marshmallow.
+        dump_to = getattr(field, 'dump_to', None)
+        load_from = getattr(field, 'load_from', None)
+        return dump_to or load_from or name
+    else:
+        return field.data_key or name
 
 
 def _get_json_type_for_field(field):
@@ -105,29 +110,69 @@ def _get_json_type_for_field(field):
     return json_type, fmt
 
 
-def field2choices(field):
-    """Return the set of valid choices for a :class:`Field <marshmallow.fields.Field>`,
-    or ``None`` if no choices are specified.
+def field2choices(field, **kwargs):
+    """Return the dictionary of swagger field attributes for valid choices definition
 
     :param Field field: A marshmallow field.
-    :rtype: set
+    :rtype: dict
     """
-    comparable = {
+    attributes = {}
+
+    comparable = [
         validator.comparable for validator in field.validators
         if hasattr(validator, 'comparable')
-    }
-    if comparable:
-        return comparable
-
-    choices = [
-        OrderedSet(validator.choices) for validator in field.validators
-        if hasattr(validator, 'choices')
     ]
-    if choices:
-        return functools.reduce(operator.and_, choices)
+    if comparable:
+        attributes['enum'] = comparable
+    else:
+        choices = [
+            OrderedSet(validator.choices) for validator in field.validators
+            if hasattr(validator, 'choices')
+        ]
+        if choices:
+            attributes['enum'] = list(functools.reduce(operator.and_, choices))
+
+    return attributes
 
 
-def field2range(field):
+def field2read_only(field, **kwargs):
+    """Return the dictionary of swagger field attributes for a dump_only field.
+
+    :param Field field: A marshmallow field.
+    :rtype: dict
+    """
+    attributes = {}
+    if field.dump_only:
+        attributes['readOnly'] = True
+    return attributes
+
+
+def field2write_only(field, **kwargs):
+    """Return the dictionary of swagger field attributes for a load_only field.
+
+    :param Field field: A marshmallow field.
+    :rtype: dict
+    """
+    attributes = {}
+    if field.load_only and kwargs['openapi_major_version'] >= 3:
+        attributes['writeOnly'] = True
+    return attributes
+
+
+def field2nullable(field, **kwargs):
+    """Return the dictionary of swagger field attributes for a nullable field.
+
+    :param Field field: A marshmallow field.
+    :rtype: dict
+    """
+    attributes = {}
+    if field.allow_none:
+        omv = kwargs['openapi_major_version']
+        attributes['x-nullable' if omv < 3 else 'nullable'] = True
+    return attributes
+
+
+def field2range(field, **kwargs):
     """Return the dictionary of swagger field attributes for a set of
     :class:`Range <marshmallow.validators.Range>` validators.
 
@@ -163,7 +208,7 @@ def field2range(field):
                 attributes['maximum'] = validator.max
     return attributes
 
-def field2length(field):
+def field2length(field, **kwargs):
     """Return the dictionary of swagger field attributes for a set of
     :class:`Length <marshmallow.validators.Length>` validators.
 
@@ -264,6 +309,12 @@ def field2property(field, spec=None, use_refs=True, dump=True, name=None):
     :param str name: The definition name, if applicable, used to construct the $ref value.
     :rtype: dict, a Property Object
     """
+    if spec:
+        openapi_major_version = spec.openapi_version.version[0]
+    else:
+        # Default to 2 for backward compatibility
+        openapi_major_version = 2
+
     from apispec.ext.marshmallow import resolve_schema_dict
     type_, fmt = _get_json_type_for_field(field)
 
@@ -274,25 +325,23 @@ def field2property(field, spec=None, use_refs=True, dump=True, name=None):
     if fmt:
         ret['format'] = fmt
 
-    default = field.default if dump else field.missing
+    default = field.missing
     if default is not marshmallow.missing:
         if callable(default):
             ret['default'] = default()
         else:
             ret['default'] = default
 
-    choices = field2choices(field)
-    if choices:
-        ret['enum'] = list(choices)
-
-    if field.dump_only:
-        ret['readOnly'] = True
-
-    if field.allow_none:
-        ret['x-nullable'] = True
-
-    ret.update(field2range(field))
-    ret.update(field2length(field))
+    for attr_func in (
+        field2choices,
+        field2read_only,
+        field2write_only,
+        field2nullable,
+        field2range,
+        field2length,
+    ):
+        ret.update(attr_func(
+            field, openapi_major_version=openapi_major_version))
 
     if isinstance(field, marshmallow.fields.Nested):
         del ret['type']
@@ -308,7 +357,7 @@ def field2property(field, spec=None, use_refs=True, dump=True, name=None):
                     raise ValueError('Must pass `name` argument for self-referencing Nested fields.')
                 # We need to use the `name` argument when the field is self-referencing and
                 # unbound (doesn't have `parent` set) because we can't access field.schema
-                ref_path = get_ref_path(spec.openapi_version.version[0])
+                ref_path = get_ref_path(openapi_major_version)
                 ref_name =  '#/{ref_path}/{name}'.format(ref_path=ref_path,
                                                          name=name)
             ref_schema = {'$ref': ref_name}
@@ -362,13 +411,12 @@ def schema2parameters(schema, **kwargs):
         raise ValueError(
             "{0!r} doesn't have either `fields` or `_declared_fields`".format(schema)
         )
-
     return fields2parameters(fields, schema, **kwargs)
 
 
 def fields2parameters(fields, schema=None, spec=None, use_refs=True,
                       default_in='body', name='body', required=False,
-                      use_instances=False):
+                      use_instances=False, description=None, **kwargs):
     """Return an array of OpenAPI parameters given a mapping between field names and
     :class:`Field <marshmallow.Field>` objects. If `default_in` is "body", then return an array
     of a single parameter; else return an array of a parameter for each included field in
@@ -385,12 +433,17 @@ def fields2parameters(fields, schema=None, spec=None, use_refs=True,
         else:
             prop = fields2jsonschema(fields, spec=spec, use_refs=use_refs, dump=False)
 
-        return [{
+        param = {
             'in': swagger_default_in,
             'required': required,
             'name': name,
             'schema': prop,
-        }]
+        }
+
+        if description:
+            param['description'] = description
+
+        return [param]
 
     assert not getattr(schema, 'many', False), \
         "Schemas with many=True are only supported for 'json' location (aka 'in: body')"