+++ tests/test_core.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from collections import OrderedDict
+
 import pytest
 import mock
 import yaml
@@ -211,6 +213,12 @@ class TestPath:
         spec.add_path(path='/path4')
         assert list(spec.to_dict()['paths'].keys()) == ['/path1', '/path2', '/path3', '/path4']
 
+    def test_methods_maintain_order(self, spec):
+        methods = ['get', 'post', 'put', 'patch', 'delete', 'head', 'options']
+        for method in methods:
+            spec.add_path(path='/path', operations=OrderedDict({method: {}}))
+        assert list(spec.to_dict()['paths']['/path']) == methods
+
     def test_add_path_merges_paths(self, spec):
         """Test that adding a second HTTP method to an existing path performs
         a merge operation instead of an overwrite"""
@@ -274,7 +282,7 @@ class TestPath:
         spec.add_path(path)
 
         p = spec._paths[path.path]
-        assert path.path == p.path
+        assert p == path.operations
         assert 'get' in p
 
     def test_add_path_strips_path_base_path(self, spec):
@@ -419,7 +427,7 @@ class TestPathHelpers:
 
     def test_path_helper_is_used(self, spec):
         def path_helper(spec, view_func, **kwargs):
-            return Path(path=view_func['path'], method='get')
+            return Path(path=view_func['path'])
         spec.register_path_helper(path_helper)
         spec.add_path(
             view_func={'path': '/pet/{petId}'},